Computer Vision/requirements.txt
# Dependências para todos os questionários de Computer Vision

# Processamento de imagens básico
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
Pillow==10.0.0

# Processamento de sinais e FFT
scipy==1.11.1

# Machine Learning (para quantização K-means)
scikit-learn==1.3.0

# Deep Learning (para Questionário 4 - VGG-16)
tensorflow==2.13.0
keras==2.13.1

# Features e matching (para Questionário 4)
opencv-contrib-python==4.8.1.78

# Utilitários
tqdm==4.65.0
seaborn==0.12.2 